refactor(mft): bring direct_index.rs and unified.rs under the 800-LOC policy for real
The previous three commits' fixes pushed both files over the file-size
policy threshold. Fixed it properly instead of adding policy exceptions:

- direct_index.rs: extracted four small helpers
  (is_primary_attribute/extract_attr_name/read_size_allocated/
  resident_and_sparse) that were each duplicated verbatim across 4-5
  match arms (Data, ReparsePoint, IndexRoot-family, ObjectId-family,
  catch-all) -- genuine deduplication, not just fewer lines. 856 -> 724
  lines, and every one of those match arms is now shorter and less
  repetitive to read, not just smaller on a line-count report.

- unified.rs: split the NTFS UTF-16/WTF-8 name-decoding cluster
  (decode_utf16le_into, decode_name_u16, wtf8_from_utf16le,
  store_name_lossless, the LOSSY_NAME_COUNT tally) into a new sibling
  module, unified/name_codec.rs. That cluster has zero dependency on
  process_record's attribute-loop state -- it only needs MftIndex -- so
  it was always a separable concern, just never separated. Re-exported
  decode_name_u16/lossy_name_count from unified.rs so the ~9 other
  modules that call them via crate::io::parser::unified::* keep working
  unchanged. 803 -> 559 lines in unified.rs; name_codec.rs is 270 lines,
  comfortably under the threshold on its own.

Also extracted a small resident_value_offset(data, attr_offset) helper
in unified.rs for the value-offset-field read duplicated between the
$FILE_NAME and $REPARSE_POINT arms.

All prior documentation is preserved verbatim -- this is a structural
split, not a comment trim. cargo test -p uffs-mft: 260/260 passing,
identical to before the split (the extraction is behavior-preserving by
construction: same logic, moved, not rewritten). Neither file needs a
scripts/ci/file_size_exceptions.txt entry anymore.
